Enum snowcap::hard_policies::LTLModal

pub enum LTLModal {
    Now(Box<dyn LTLOperator>),
    Next(Box<dyn LTLOperator>),
    Finally(Box<dyn LTLOperator>),
    Globally(Box<dyn LTLOperator>),
    Until(Box<dyn LTLOperator>, Box<dyn LTLOperator>),
    Release(Box<dyn LTLOperator>, Box<dyn LTLOperator>),
    WeakUntil(Box<dyn LTLOperator>, Box<dyn LTLOperator>),
    StrongRelease(Box<dyn LTLOperator>, Box<dyn LTLOperator>),
}

Temporal modal operators of LTL. For reconfiguration purpose, in the last state, we assume that nothing changes anymore, and every propositional variable does not change its state. See here

As an example, let $\phi_1$ be the policy that needs to be satisfied initially, and $\phi_2$ the policy that needs to be satisfied at the end. Additionally, we would like that we switch once from $\phi_1$ to $\phi_2$. This would be the following expression:

$$\phi_1\ \mathbf{U}\ \mathbf{G}\ \phi_2$$

Variants

Now(Box<dyn LTLOperator>)

$\phi$: $\phi$ holds at the current state.

Next(Box<dyn LTLOperator>)

$\mathbf{X}\ \phi$: $\phi$ holds at the next state. If the sequence is finished, then $\mathbf{X} \phi \iff \phi$ is identical to stating that $\phi$ holds now.

Finally(Box<dyn LTLOperator>)

$\mathbf{F}\ \phi$: $\phi$ needs to hold eventually (once)

Globally(Box<dyn LTLOperator>)

$\mathbf{G}\ \phi$: $\phi$ needs to hold in the current and every future state

Until(Box<dyn LTLOperator>, Box<dyn LTLOperator>)

$\psi\ \mathbf{U}\ \phi$: $\psi$ has to hold at least until (but not including the state where) $\phi$ becomes true. $\phi$ can hold at the current or a future position (once). Until then, $psi$ must be true. $\phi$ must hold eventually!

Release(Box<dyn LTLOperator>, Box<dyn LTLOperator>)

$\psi\ \mathbf{R}\ \phi$: $\phi$ has to hold until and including the point where $\psi$ first holds. If $\psi$ never holds, then $\phi$ must hold forever.

WeakUntil(Box<dyn LTLOperator>, Box<dyn LTLOperator>)

$\psi\ \mathbf{W}\ \phi$: $\psi$ has to hold at least at least util (but not including the state where) $\phi$ becomes true. If $\phi$ never holds, $\psi$ must hold forever.

StrongRelease(Box<dyn LTLOperator>, Box<dyn LTLOperator>)

$\psi\ \mathbf{M}\ \phi$: $\phi$ has to hold until and including the point where $\psi$ first holds. $\psi$ can hold now or at any future state, but $\psi$ must hold eventually!