Struct snowcap::hard_policies::HardPolicy

pub struct HardPolicy {
    pub prop_vars: Vec<Condition>,
    pub expr: LTLModal,
    // some fields omitted
}

Linear Temporal Logic

This structure holds the entire LTL expression. It is stored as a vector of propositional a history of which constraints were satisfied, and an expression which can check the property based on the history.

Fields
prop_vars: Vec<Condition>

Conditional variables of the hard poicy

expr: LTLModal

LTL Expression

Implementations

impl HardPolicy

pub fn reachability<'r, 'p, R, P>(routers: R, prefixes: P) -> Self where
R: Iterator<Item = &'r RouterId> + Clone,
P: Iterator<Item = &'p Prefix> + Clone,

Helper function to generate the reachability policy

pub fn globally(prop_vars: Vec<Condition>) -> Self

Create a new Linear Temporal Logic Hard Policy, where all conditions supplied need to be satisfied all the time.

pub fn until_globally(
prop_vars: Vec<Condition>,
phi: &[usize],
psi: &[usize]
) -> Self

Create a new Linear Temporal Logic Hard Policy. The LTL expression will have the following form, where $\phi_i$ are all propositional variables in the second argument, and $\psi_i$ are all from the third argument.

$$ \Big( \bigwedge \phi_i \Big) \ \mathbf{U}\ \mathbf{G}\ \Big( \bigwedge \psi_i \Big) $$

pub fn new(prop_vars: Vec<Condition>, expr: LTLModal) -> Self

Create a new Linear Temporal Logic Hard Policy

pub fn set_num_mods_if_none(&mut self, num_mods: usize) -> bool

Sets the total number of modifiers, if it was not yet set before. If it is already set, then nothing will change. This function returns true if there was no previous value.

pub fn step(
&mut self,
net: &mut Network,
state: &mut ForwardingState
) -> Result<(), NetworkError>

Applies a next step to the LTL model

pub fn undo(&mut self)

Undoes the last call to step

pub fn reset(&mut self)

Reset the strucutre, such that no state exists.

pub fn check(&self) -> bool

Computes the condition using Linear Temporal Logic, and returns if the expression holds. The finish flag will be determined automatically if the number of calls to step subtracted by the number of calls to undo (i.e., the length of the history) is equal to the number of modifiers plus one (since the initial state will also be part of the history.) Finish will be set to false if the number of modifiers was not set yet.

pub fn check_overwrite_finish(&self, finish: bool) -> bool

Computes the condition using Linear Temporal Logic, and returns if the expression holds. The provided finish flag will be used to determine the method (check vs partial).

pub fn get_watch(&self) -> Vec<usize>

Compute the set of propositional variables that need to be watched in order to change the outcome of the current state of the checker. This function should only be used when the result is either false or undefined.

pub fn get_watch_errors(&self) -> WatchErrors

Compute the watch and get the last errors of the respective watch. This is a helpful utility for getting the errors, and comparing it later.

pub fn get_watch_overwrite_finish(&self, finish: bool) -> Vec<usize>

Compute the set of propositional variables that need to be watched in order to change the outcome of the current state of the checker. This function should only be used when the result is either false or undefined. The finish flagg will be used to determine the method (watch or watch_partial).

pub fn last_errors(&self) -> HashSet<PolicyError>

Returns the set of all errors of all propositional variables. It might be the case that these errors don't necessarily contribute to the result of the last call to check.

pub fn last_errors_of_watch(&self, watch: &[usize]) -> Vec<Option<PolicyError>>

Get the error associated with the watch provided to this method. The watch is an array containing the indices of all propositional variables which need to be considered. The resulting vector contains the error of the propositional variable, in the same order as provided. If a propositional variable is true, the error corresponding to this variable is None.

pub fn compare_watch_errors(&self, watch_errors: &WatchErrors) -> bool

This method compares the current state of the checker with a previous state, which was extracted using the method get_watch_errors. This funciton returns true if the errors are the same, and false if the errors are different. Only the errors from the watch are compared.

pub fn repr_with_name(&self, net: &Network) -> String

Represent the LTL condition by a multiline string