Trait snowcap::hard_policies::LTLOperator

pub trait LTLOperator: Debug + LTLBoxClone + Send + UnwindSafe + RefUnwindSafe {
    fn check(&self, history: &[Vec<bool>]) -> bool;

    fn partial(&self, history: &[Vec<bool>]) -> LTLResult;

    fn watch(&self, history: &[Vec<bool>]) -> Vec<usize>;

    fn watch_partial(&self, history: &[Vec<bool>]) -> Vec<usize>;

    fn repr(&self) -> String;
}

Operators of LTL

An operator may either be a simple propositional variable, a boolean or a temporal modal operator.

Required methods

fn check(&self, history: &[Vec<bool>]) -> bool

Checks if the operator holds for the given history. For this, it is assumed that the sequence is finished!

fn partial(&self, history: &[Vec<bool>]) -> LTLResult

Checks if the operator holds for the given history, assuming that we have only a partial sequence.

fn watch(&self, history: &[Vec<bool>]) -> Vec<usize>

Extract the set of propositional variables, that need to change in order for the result of the operator to change. Note, that not all variables in the returned set need to change, but only a subset. Some examples:

use snowcap::hard_policies::*;
use snowcap_ltl_parser::ltl;

assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![true, true, true]]), vec![0, 1, 2]);
assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![true, true, false]]), vec![2]);
assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![true, false, false]]), vec![1, 2]);
assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![false, false, false]]), vec![0, 1, 2]);

fn watch_partial(&self, history: &[Vec<bool>]) -> Vec<usize>

Extract the set of propositional variables, that need to change in order for the result of the operator to change. Note, that not all variables in the returned set need to change, but only a subset. For this funciton, the history is not yet complete. Some examples:

use snowcap::hard_policies::*;
use snowcap_ltl_parser::ltl;

assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![true, true, true]]), vec![0, 1, 2]);
assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![true, true, false]]), vec![2]);
assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![true, false, false]]), vec![1, 2]);
assert_eq!(ltl!(And(0, 1, 2)).watch(&vec![vec![false, false, false]]), vec![0, 1, 2]);

fn repr(&self) -> String

represent the operator as a string

Loading content...