Module snowcap::netsim::config

Network Configuration

This module represents the network configuration. There are several different structs in this module. Here is an overview:

Config: Network-wide configuration. The datastructure is a collection of several ConfigExpr.

ConfigExpr: Single configuration expresison (line in a router configuraiton).

ConfigPatch: Difference between two Config structs. The datastructure is a collection of several ConfigModifier.

ConfigModifier: A modification of a single ConfigExpr in a configuration. A modification can either be an insertion of a new expression, a removal of an existing expression, or a moification of an existing expression.

Example Usage

use snowcap::netsim::BgpSessionType::*;
use snowcap::netsim::config::{Config, ConfigExpr::BgpSession, ConfigModifier};
use snowcap::netsim::ConfigError;

fn main() -> Result<(), ConfigError> {
    // routers
    let r0 = 0.into();
    let r1 = 1.into();
    let r2 = 2.into();
    let r3 = 3.into();
    let r4 = 4.into();

    let mut c1 = Config::new();
    let mut c2 = Config::new();

    // add the same bgp expression
    c1.add(BgpSession { source: r0, target: r1, session_type: IBgpPeer })?;
    c2.add(BgpSession { source: r0, target: r1, session_type: IBgpPeer })?;

    // add one only to c1
    c1.add(BgpSession { source: r0, target: r2, session_type: IBgpPeer })?;

    // add one only to c2
    c2.add(BgpSession { source: r0, target: r3, session_type: IBgpPeer })?;

    // add one to both, but differently
    c1.add(BgpSession { source: r0, target: r4, session_type: IBgpPeer })?;
    c2.add(BgpSession { source: r0, target: r4, session_type: IBgpClient })?;

    // Compute the patch (difference between c1 and c2)
    let patch = c1.get_diff(&c2);
    // Apply the patch to c1
    c1.apply_patch(&patch)?;
    // c1 should now be equal to c2
    assert_eq!(c1, c2);

    Ok(())
}

Structs

Config
Network Configuration

ConfigPatch
Config Patch

Enums

ConfigExpr
Single configuration expression

ConfigExprKey
Key for Config Expressions

ConfigModifier
Config Modifier