Enum snowcap::netsim::config::ConfigExpr

pub enum ConfigExpr {
    IgpLinkWeight {
        source: RouterId,
        target: RouterId,
        weight: LinkWeight,
    },
    BgpSession {
        source: RouterId,
        target: RouterId,
        session_type: BgpSessionType,
    },
    BgpRouteMap {
        router: RouterId,
        direction: RouteMapDirection,
        map: RouteMap,
    },
    StaticRoute {
        router: RouterId,
        prefix: Prefix,
        target: RouterId,
    },
}

Single configuration expression

The expression sets a specific thing in the network.

Variants

IgpLinkWeight

Sets the link weight of a single link (directional) TODO make sure that the weight is strictly smaller than infinity.

Fields of IgpLinkWeight

source: RouterId

Source router for link

target: RouterId

Target router for link

weight: LinkWeight

Link weight for IGP

BgpSession

Create a BGP session TODO currently, this is treated as a single configuration line, where in fact, it is two distinct configurations, one on the source and one on the target. We treat it as a single configuration statement, because it is only active once both speakers have opened the session. Changing this requires changes in router.rs.

Fields of BgpSession

source: RouterId

Source router for Session

target: RouterId

Target router for Session

session_type: BgpSessionType

Session type

BgpRouteMap

Set the BGP Route Map

Fields of BgpRouteMap

router: RouterId

Router to configure the route map

direction: RouteMapDirection

Direction (incoming or outgoing)

map: RouteMap

Route Map

StaticRoute

Set a static route

Fields of StaticRoute

router: RouterId

On which router set the static route

prefix: Prefix

For which prefix to set the static route

target: RouterId

To which neighbor to forward packets to.

Implementations

impl ConfigExpr

pub fn key(&self) -> ConfigExprKey

Returns the key of the config expression. The idea behind the key is that the ConfigExpr cannot be hashed and used as a key for a HashMap. But ConfigExprKey implements Hash, and can therefore be used as a key.

pub fn routers(&self) -> Vec<RouterId>

Returns the router IDs on which the configuration is applied and have to be changed.