Enum snowcap::netsim::NetworkDevice

pub enum NetworkDevice<'a> {
    InternalRouter(&'a Router),
    ExternalRouter(&'a ExternalRouter),
    None,
}

Network Device (similar to Option)

Enumerates all possible network devices. This struct behaves similar to an Option, but it knows two different Some values, the InternalRouter and the ExternalRouter. Thus, it knows three different unwrap functions, the unwrap_internal, unwrap_external and unwrap_none function, as well as internal_or and external_or.

Variants

InternalRouter(&'a Router)

Internal Router

ExternalRouter(&'a ExternalRouter)

External Router

None

None was found

Implementations

impl<'a> NetworkDevice<'a>

pub fn unwrap_internal(self) -> &'a Router

Returns the Router or panics, if the enum is not a NetworkDevice::InternalRouter

pub fn unwrap_external(self) -> &'a ExternalRouter

Returns the Router or panics, if the enum is not a NetworkDevice::ExternalRouter

pub fn unwrap_none(self)

Returns () or panics is the enum is not a NetworkDevice::None

pub fn is_internal(&self) -> bool

Returns true if and only if self contains an internal router.

pub fn is_external(&self) -> bool

Returns true if and only if self contains an external router.

pub fn is_none(&self) -> bool

Returns true if and only if self contains NetworkDevice::None.

pub fn internal(self) -> Option<&'a Router>

Maps the NetworkDevice to an option, with Some(r) only if self is InternalRouter.

pub fn external(self) -> Option<&'a ExternalRouter>

Maps the NetworkDevice to an option, with Some(r) only if self is ExternalRouter.

pub fn internal_or<E: Error>(self, error: E) -> Result<&'a Router, E>

Maps the NetworkDevice to result, with the Ok case only if self is InternalRouter. If self is not InternalError, then the provided error is returned.

pub fn external_or<E: Error>(self, error: E) -> Result<&'a ExternalRouter, E>

Maps the NetworkDevice to result, with the Ok case only if self is ExternalRouter. If self is not ExternalRouter, then the provided error is returned.

pub fn none_or<E: Error>(self, error: E) -> Result<(), E>

Maps the NetworkDevice to result, with the Ok case only if self is none. If self is not None, then the provided error is returned.