Enum snowcap::netsim::NetworkError

pub enum NetworkError {
    DeviceError(DeviceError),
    ConfigError(ConfigError),
    DeviceNotFound(RouterId),
    DeviceNameNotFound(String),
    DeviceIsExternalRouter(RouterId),
    ForwardingLoop(Vec<RouterId>),
    ForwardingBlackHole(Vec<RouterId>),
    InvalidBgpSessionType(RouterId, RouterId, BgpSessionType),
    ConvergenceLoop(Vec<Event>, Vec<Network>),
    NoConvergence,
    RoutersNotConnected(RouterId, RouterId),
    InvalidBgpTable(RouterId),
    UnexpectedEventConvergenceLoop,
    InvalidEvent(Event),
    HistoryError(&'static str),
    UnsatisfiedConstraints(PolicyError),
    NoEventsToReorder,
}

Network Errors

Variants

DeviceError(DeviceError)

Device Error which cannot be handled

ConfigError(ConfigError)

Configuration error

DeviceNotFound(RouterId)

Device is not present in the topology

DeviceNameNotFound(String)

Device name is not present in the topology

DeviceIsExternalRouter(RouterId)

Device must be an internal router, but an external router was passed

ForwardingLoop(Vec<RouterId>)

Forwarding loop detected

ForwardingBlackHole(Vec<RouterId>)

Black hole detected

InvalidBgpSessionType(RouterId, RouterId, BgpSessionType)

Invalid BGP session type

ConvergenceLoop(Vec<Event>, Vec<Network>)

Convergence Problem, but loop was detected

NoConvergence

Convergence Problem

RoutersNotConnected(RouterId, RouterId)

Two routers are not adjacent

InvalidBgpTable(RouterId)

The BGP table is invalid

UnexpectedEventConvergenceLoop

Error encountered while finding convergence loop. Enqueued event does not match the expectation.

InvalidEvent(Event)

Event cannot be handled by the network

HistoryError(&'static str)

History is invalid

UnsatisfiedConstraints(PolicyError)

Constraints are not satisfied during convergence

NoEventsToReorder

No events to reorder (This error is only thorwn when approximating transient state violation probability in Network::apply_modifier_check_transient)