Module snowcap::netsim::printer

Helper (printer) functions for the Network

Module containing helper functions to get formatted strings and print information about the network.

Functions

bgp_entry
Returns the formatted string for a given RIBEntry of a BGP route. The entry is formatted such that all router names are inserted.

bgp_route
Returns a formatted string for a given BGP route.

bgp_table
Get a vector of strings, which represent the bgp table. Each String in the vector represents one line (one known route). The strings are formatted, and the names of the routers are inserted. The chosen route is prefixed with a *, while the known but not chosen routes are prefixed with nothing ( ).

config_expr
Returns the config expr as a string, where all router names are inserted.

config_modifier
Returns a formatted string for the given modifier, where all router names are inserted.

event
Return a formatted string for a given event

print_bgp_table
Print the bgp table of a given router.

print_config
Print the complete configuration to stdout

print_config_patch
Print the configuration patch to stdout

route_map
Returns a formatted string of the route map, where all router names are inserted