Module snowcap::netsim::route_map

Route-Maps

This module contains the necessary structures to build route maps for internal BGP routers.

Structs

RouteMap
Main RouteMap structure

RouteMapBuilder
Route Map Builder

Enums

RouteMapDirection
Direction of the Route Map

RouteMapMatch
Match statement of the route map. Can be combined to generate complex match statements

RouteMapMatchAsPath
Clause to match on the as path

RouteMapMatchClause
Generic RouteMapMatchClause to match on all, a range or on a specific element

RouteMapSet
Set action, if a route map matches

RouteMapState
State of a route map, which can either be allow or deny