Struct snowcap::netsim::route_map::RouteMapBuilder

pub struct RouteMapBuilder { /* fields omitted */ }

Route Map Builder

Convenience type to build a route map. You are required to at least call order and state once on the builder, before you can call build. If you don't call add_match (or any function adding a match statement) on the builder, it will match on any route.

let map = RouteMapBuilder::new()
    .order(10)
    .allow()
    .match_neighbor(neighbor)
    .match_prefix(prefix)
    .set_community(1)
    .reset_local_pref()
    .build();

Implementations

impl RouteMapBuilder

pub fn new() -> Self

Create an empty RouteMapBuilder

pub fn order(&mut self, order: usize) -> &mut Self

Set the order of the Route-Map.

pub fn state(&mut self, state: RouteMapState) -> &mut Self

Set the state of the Route-Map.

pub fn allow(&mut self) -> &mut Self

Set the state of the Route-Map to allow. This function is identical to calling state(RouteMapState::Allow).

pub fn deny(&mut self) -> &mut Self

Set the state of the Route-Map to deny. This function is identical to calling state(RouteMapState::Deny).

pub fn cond(&mut self, cond: RouteMapMatch) -> &mut Self

Add a match condition to the Route-Map.

pub fn match_neighbor(&mut self, neighbor: RouterId) -> &mut Self

Add a match condition to the Route-Map, matching on the neighbor

pub fn match_prefix(&mut self, prefix: Prefix) -> &mut Self

Add a match condition to the Route-Map, matching on the prefix with exact value

pub fn match_prefix_range(&mut self, from: Prefix, to: Prefix) -> &mut Self

Add a match condition to the Route-Map, matching on the prefix with an inclusive range

pub fn match_as_path_contains(&mut self, as_id: AsId) -> &mut Self

Add a match condition to the Route-Map, requiring that the as path contains a specific AS

pub fn match_as_path_length(&mut self, as_path_len: usize) -> &mut Self

Add a match condition to the Route-Map, matching on the as path length with exact value

pub fn match_as_path_length_range(
&mut self,
from: usize,
to: usize
) -> &mut Self

Add a match condition to the Route-Map, matching on the as path length with an inclusive range

pub fn match_next_hop(&mut self, next_hop: RouterId) -> &mut Self

Add a match condition to the Route-Map, matching on the next hop

pub fn match_community(&mut self, community: u32) -> &mut Self

Add a match condition to the Route-Map, matching on the community with exact value

pub fn match_community_empty(&mut self) -> &mut Self

Add a match condition to the Route-Map, matching on routes without community

pub fn match_community_range(&mut self, from: u32, to: u32) -> &mut Self

Add a match condition to the Route-Map, matching on the community with an inclusive range

pub fn add_set(&mut self, set: RouteMapSet) -> &mut Self

Add a set expression to the Route-Map.

pub fn set_next_hop(&mut self, next_hop: RouterId) -> &mut Self

Add a set expression, overwriting the next hop value

pub fn set_local_pref(&mut self, local_pref: u32) -> &mut Self

Add a set expression, overwriting the Local-Pref

pub fn reset_local_pref(&mut self) -> &mut Self

Add a set expression, resetting the local-pref

pub fn set_med(&mut self, med: u32) -> &mut Self

Add a set expression, overwriting the MED

pub fn reset_med(&mut self) -> &mut Self

Add a set expression, resetting the MED

pub fn set_igp_cost(&mut self, cost: LinkWeight) -> &mut Self

Add a set expression, overwriting the Igp Cost to reach the next-hop

pub fn set_community(&mut self, community: u32) -> &mut Self

Add a set expression, overwriting the Community

pub fn reset_community(&mut self) -> &mut Self

Add a set expression, resetting the Community

pub fn build(&self) -> RouteMap

Build the route-map.

Panics

The function panics in the following cases:

The order is not set (order was not called),

The state is not set (neither state, allow nor deny were called),