Trait snowcap::permutators::Permutator

pub trait Permutator<T> where
    Self: Iterator,
    Self::Item: PermutatorItem<T>, {
    fn new(input: Vec<T>) -> Self;

    fn fail_pos(&mut self, _pos: usize) { ... }
}

Permutator trait

Required methods

fn new(input: Vec<T>) -> Self

Creates the permutator from the sequence of T

Loading content...
Provided methods

fn fail_pos(&mut self, _pos: usize)

This function may reduce the number of permutations, by skipping permutations which start the exact same way as the last call to next, up to the position pos. Not every permutator has this funciton implemented.

Loading content...