Trait snowcap::permutators::PermutatorItem

pub trait PermutatorItem<T> {
    fn as_patches(self) -> Vec<T>;
}

This is an empty trait to tell the compiler which types can be returned by the Permutator Iterator.

Required methods

fn as_patches(self) -> Vec<T>

Transforms the item to a vector of T. This is in order to allow a Permutator to be passed to a function dynamically.

Loading content...