Trait snowcap::soft_policies::SoftPolicy

pub trait SoftPolicy {
    fn new(state: &mut ForwardingState, net: &Network) -> Self;

    fn update(&mut self, state: &mut ForwardingState, net: &Network);

    fn cost(&self) -> f64;
}

Trait for creating soft policies.

Required methods

fn new(state: &mut ForwardingState, net: &Network) -> Self

Crates a new soft policy and initializes it with correct initial values.

fn update(&mut self, state: &mut ForwardingState, net: &Network)

Update the information in the SoftPolicy. This function must be called after every modifier is applied.

fn cost(&self) -> f64

Compute the score based on the information gathered by several calls to update. The output of this funciton is between 0 and 1, and lower is better.

Loading content...