Trait snowcap::strategies::Strategy

pub trait Strategy {
    fn new(
        net: Network, 
        modifiers: Vec<ConfigModifier>, 
        hard_policy: HardPolicy, 
        time_budget: Option<Duration>
    ) -> Result<Box<Self>, Error>;

    fn work(&mut self, abort: Stopper) -> Result<Vec<ConfigModifier>, Error>;

    fn num_states(&self) -> usize;

    fn synthesize(
        net: Network, 
        end_config: Config, 
        hard_policy: HardPolicy, 
        time_budget: Option<Duration>, 
        abort: Stopper
    ) -> Result<Vec<ConfigModifier>, Error> { ... }
}

Infterface for all strategies

Required methods

fn new(
net: Network,
modifiers: Vec<ConfigModifier>,
hard_policy: HardPolicy,
time_budget: Option<Duration>
) -> Result<Box<Self>, Error>

Create the strategy

fn work(&mut self, abort: Stopper) -> Result<Vec<ConfigModifier>, Error>

Main function to find a valid reconfiguration sequence (if it exists) and return it. The function also returns the number of sequences that were tested.

fn num_states(&self) -> usize

Returns the number of states explored by the strategy.

This method is only available if the "count-states" feature is enabled!

Loading content...
Provided methods

fn synthesize(
net: Network,
end_config: Config,
hard_policy: HardPolicy,
time_budget: Option<Duration>,
abort: Stopper
) -> Result<Vec<ConfigModifier>, Error>

Wrapper, that creates the strategy and synthesizes the network update order.

Loading content...