Enum snowcap::topology_zoo::Scenario

pub enum Scenario {
    FullMesh2RouteReflector,
    RouteReflector2FullMesh,
    DoubleIgpWeight,
    HalveIgpWeight,
    IntroduceSecondRouteReflector,
    RemoveSecondRouteReflector,
    NetworkAcquisition,
    NetworkSplit,
    DisconnectRouter,
    ConnectRouter,
    DoubleLocalPref,
    HalveLocalPref,
    VerifyTransientCondition,
    VerifyTransientConditionReverse,
}

Scenario to apply to a topology in TopologyZoo

Variants

FullMesh2RouteReflector

Scenario, where we start with a iBGP full mesh, and end up with a topology, where one single router is elected as a Route Reflectors, and all others pair with that router.

RouteReflector2FullMesh

Scenario, where we start with a topology, where one single router is elected as a Route Reflectors, and all others pair with that router, and we end up wiht an iBGP full mesh.

DoubleIgpWeight

Scenario, where every IGP weight is doubled

HalveIgpWeight

Scenario, where every IGP weight is halved

IntroduceSecondRouteReflector

Scenario, where we start with a single Route-Reflector, to which all other routers pair, and end with a second Route-Reflector as a backup, where all other routers have a session to both reflectors, and the two reflectors are connected with a peer.

RemoveSecondRouteReflector

Scenario, where we start with a second Route-Reflector as a backup, where all other routers have a session to both reflectors, and the two reflectors are connected with a peer, and end with a single Route-Reflector, to which all other routers pair.

NetworkAcquisition

Scenario, where we start with two different connected components, both having connection to the outside world, and we merge them by adding the links in between.

NetworkSplit

Reverse scenario of the Network Acquisition

DisconnectRouter

Shutdown a router by setting all link weights around a specific router to infinity

ConnectRouter

Introduce a new router by starting with infinite weight of all links around a specific router, and setting them to something reasonable.

DoubleLocalPref

Scenario where every local-pref configuration is doulbed

HalveLocalPref

Scenario where every local-pref is halved

VerifyTransientCondition

Test scenario for verifying transient state conditions. This scenario contains only a single modifier, which adds an eBGP session.

VerifyTransientConditionReverse

Test scenario for verifying transient state conditions. This scenario contains only a single modifier, which adds an eBGP session.